adjust test cases
