+++ dubbo-remoting/dubbo-remoting-netty4/src/main/resources/META-INF/dubbo/internal/org.apache.dubbo.remoting.Codec2
@@ -1,0 +1,2 @@
+tri=org.apache.dubbo.remoting.transport.netty4.DefaultCodec
+empty=org.apache.dubbo.remoting.transport.netty4.DefaultCodec
