+++ dubbo-remoting/dubbo-remoting-netty4/src/main/resources/META-INF/dubbo/internal/org.apache.dubbo.remoting.api.WireProtocol
@@ -1,0 +1,1 @@
+empty=org.apache.dubbo.remoting.transport.netty4.EmptyWireProtocol
