+++ dubbo-remoting/dubbo-remoting-netty4/src/test/java/org/apache/dubbo/remoting/transport/netty4/PortUnificationExchangerTest.java
@@ -1,0 +1,43 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.dubbo.remoting.transport.netty4;
+
+import org.apache.dubbo.common.URL;
+import org.apache.dubbo.common.constants.CommonConstants;
+import org.apache.dubbo.common.url.component.ServiceConfigURL;
+import org.apache.dubbo.remoting.Constants;
+import org.apache.dubbo.remoting.api.pu.DefaultPuHandler;
+import org.apache.dubbo.remoting.exchange.PortUnificationExchanger;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+
+public class PortUnificationExchangerTest {
+
+    @Test
+    public void test() {
+        URL url = new ServiceConfigURL(CommonConstants.TRIPLE, "localhost", 9103,
+            new String[]{Constants.BIND_PORT_KEY, String.valueOf(9103)});
+        PortUnificationExchanger.bind(url, new DefaultPuHandler());
+        PortUnificationExchanger.bind(url, new DefaultPuHandler());
+        Assertions.assertEquals(PortUnificationExchanger.getServers().size(), 1);
+
+        PortUnificationExchanger.close();
+        Assertions.assertEquals(PortUnificationExchanger.getServers().size(), 0);
+    }
+
+}
