+++ dubbo-remoting/dubbo-remoting-netty4/src/test/java/org/apache/dubbo/remoting/transport/netty4/PortUnificationServerTest.java
@@ -14,28 +14,28 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.dubbo.remoting.exchange;
+package org.apache.dubbo.remoting.transport.netty4;
 
 import org.apache.dubbo.common.URL;
 import org.apache.dubbo.common.constants.CommonConstants;
 import org.apache.dubbo.common.url.component.ServiceConfigURL;
 import org.apache.dubbo.remoting.Constants;
+import org.apache.dubbo.remoting.RemotingException;
+import org.apache.dubbo.remoting.api.pu.DefaultPuHandler;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-public class PortUnificationExchangerTest {
+public class PortUnificationServerTest {
 
     @Test
-    public void test() {
-        URL url = new ServiceConfigURL(CommonConstants.TRIPLE, "localhost", 9103,
-            new String[]{Constants.BIND_PORT_KEY, String.valueOf(9103)});
-        PortUnificationExchanger.bind(url);
-        PortUnificationExchanger.bind(url);
-        Assertions.assertEquals(PortUnificationExchanger.getServers().size(), 1);
+    public void testBind() throws RemotingException {
+        URL url = new ServiceConfigURL(CommonConstants.TRIPLE, "localhost", 8898,
+                new String[]{Constants.BIND_PORT_KEY, String.valueOf(8898)});
 
-        PortUnificationExchanger.close();
-        Assertions.assertEquals(PortUnificationExchanger.getServers().size(), 0);
+        // abstract endpoint need to get codec of url(which is in triple package)
+        final NettyPortUnificationServer server = new NettyPortUnificationServer(url, new DefaultPuHandler());
+        server.bind();
+        Assertions.assertTrue(server.isBound());
     }
-
 }
